+++ NEWS.md
@@ -3,7 +3,7 @@
 ## MsFeatures 0.0.3
 
 - Add functionality to group features based on similar retention times.
-- Rename *old* `groupClosest` function to `groupSorted`.
+- Rename *old* `groupClosest` function to `groupConsecutive`.
 - Add new `groupClosest` function which uses `groupSimilarityMatrix`.
 
 ## MsFeatures 0.0.2